+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated Lua version to 5.4
 - Use Lua path environment variables specific for version 5.4
+- Move uarch halt from assembler to C++
 
 ## [0.14.0] - 2023-05-03
 ### Added